+++ config/initializers/summit_requests.rb
@@ -1,0 +1,1 @@
+require 'summit_requests'
